Revert "merge yaml by flatting and refactor code"
This reverts commit b07726df0435534350f415748c4d38bb3cb73fd0.
